refactor(frontend): deduplicate smart alert hooks to SSOT (Phase 2)
